Added scripts that allow working with the rte-rttmgp repo
The 0001 patch must be applied ontop of the rte-rttmgp repo for now, A
temporary measure, It allows us to compile the rte-rttmgp as a shared
library that we can then use to bind to python

The compile_fortran.sh script compiles the fortran code into a shared
library that we can then use.

Really basic usage intructions:

1. Apply 0001-...-w.patch into rte-rrtmgp submodule
2. Run compile_fortran.sh (check if you need to install dependencies on
   top of script)
3. Run run.sh (this will compile our python bindings to a folder called
   rttpy and will copy test.py to the same folder
4. cd rttpy
5. python3 test.py

We expect to see 2 TEST_PASSED prints and then 2 arrays of 10 elements
printed, First is of random values, second should be zeros.

The test invokes a fortran method called zero 1d array and runs 3 tests
on this method