examples -> Jupyterbook (#247)
Converts the files in the top level of examples/ into a Jupyterbook v2. The files are converted to py:percent to run the CI but are more human-readable as a Jupyter book. The book is published on Github Pages Instructions are provided so people can convert the markdown to interactive Jupyter Notebooks if they want to experiment.